backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
transformers==4.36.0
torch==2.1.2
chromadb==0.4.23
google-generativeai==0.5.4
sentence-transformers==2.3.1
scikit-learn==1.3.2
numpy==1.24.3
pdfplumber==0.11.7
python-docx==1.2.0
